Fix GCC 15 build: cstdint includes, CURLMSG format, shuffle
- Include <cstdint> in curl.hpp and download_result.hpp for uint8_t
- Cast CURLMSG to int for std::format in download_manager
- Replace deprecated std::random_shuffle with std::shuffle

+++ src/network/download_manager.cpp
@@ -168,7 +168,7 @@ DownloadManager::process_curl_data()
         break;
 
       default:
-        log_error("unhandled cURL message: {}", msg->msg);
+        log_error("unhandled cURL message: {}", static_cast<int>(msg->msg));
         break;
     }
   }
